Extract a new TextEditor component from the job post modal and rearange the component folder.
